// src/interface/AddEditUserProfile.interface.ts
export interface AddEditUserProfilePersonalInfoInterface {
  first_name: string;
  middle_name: string;
  last_name: string;
  full_name: string;
  profile_picture: string;
  gender: string;
  date_of_birth: null | Date;
  blood_group: string;
  about: string;
}

export interface AddEditUserProfileEmployeeInfoInterface {
  status: string;
  organization_name: string;
  employee_code: string;
  department: string;
  designation: string;
  employee_email: string;
  reporting_to: { id: string; name: string };
  employee_role: { role_id: string; role_name: string };
  employee_type: string;
  joining_date: null | Date;
}

export interface AddEditUserProfilePersonalContactInfo {
  personal_email: string;
  mobile_number: string;
  country_info: string;
  emergency_contacts: {
    emergency_contact_country_info: string;
    emergency_contact_number: string;
    emergency_contact_name: string;
    contact_id: string;
    id: string;
  }[];
}

export interface AddEditUserFamilyInfo {
  father_name: string;
  mother_name: string;
  marital_status: string;
  children: {
    child_name: string;
    child_date_of_birth: null | Date;
    family_info_id: string;
    id: string;
  }[];
}

export interface AddEditUserCurrentAddress {
  address: string;
  country: string;
  city: string;
  state: string;
  zip_code: string;
  country_code: string;
}

export interface AddEditUserSocialLinks {
  icon: string;
  name: string;
  link: string;
  target_blank: boolean;
  id: string;
  user_id: string;
}

export interface AddEditUserProfileInterFace {
  personal_info: AddEditUserProfilePersonalInfoInterface;
  employee_info: AddEditUserProfileEmployeeInfoInterface;
  personal_contact_info: AddEditUserProfilePersonalContactInfo;
  family_info: AddEditUserFamilyInfo;
  current_address: AddEditUserCurrentAddress;
  same_as_current_address: boolean;
  permanent_address: AddEditUserCurrentAddress;
  social_link: AddEditUserSocialLinks[];
}
// ...
export interface EmployeeAddressInfoInterface {
  formData: AddEditUserProfileInterFace;
  setFormData: React.Dispatch<SetStateAction<AddEditUserProfileInterFace>>;
  showEmptyFieldError: boolean;
  formSubmitLoader: boolean;
  isFetchingCountryData: boolean;
  countryData: CountryDataInterface[];
  selectedCountryInfoForCurrentAddress: CountryDataInterface;
  setSelectedCountryInfoForCurrentAddress: React.Dispatch<
    SetStateAction<CountryDataInterface>
  >;
  handleOnChange: (
    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
  ) => void;
  disabled: boolean;
}